+++ front/src/pages/FaqPage.js
@@ -17,7 +17,6 @@ const Faq = () => {
   const [open1, setOpen1] = useState(false);
   const [open2, setOpen2] = useState(false);
   const [open3, setOpen3] = useState(false);
-  const [open4, setOpen4] = useState(false);
 
   return (
     <>
