+++ front/src/pages/TemplatePage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const TemplatePage = () => {
+  return <div></div>;
+};
+
+export default TemplatePage;
